Refactor db.ts into modular db/ directory
Split the 571-line db.ts file into smaller, focused modules:
- db/client.ts - database connection management
- db/migrations/ - table creation and migrations
- db/settings.ts - settings table operations
- db/events.ts - events table operations
- db/attendees.ts - attendees table operations
- db/sessions.ts - sessions table operations
- db/login-attempts.ts - rate limiting operations
- db/index.ts - re-exports for API compatibility

All imports updated to use the new module structure.
