+++ package.json
@@ -8,6 +8,8 @@
     "#fp/*": "./src/fp/*",
     "#test-utils": "./src/test-utils/index.ts",
     "#test-utils/*": "./src/test-utils/*",
+    "#lib/db": "./src/lib/db/index.ts",
+    "#lib/db/*": "./src/lib/db/*",
     "#lib/*": "./src/lib/*",
     "#routes": "./src/routes/index.ts",
     "#routes/*": "./src/routes/*",
