+++ src/edge/bunny-script.ts
@@ -6,7 +6,7 @@
 import * as BunnySDK from "@bunny.net/edgescript-sdk";
 import { once } from "#fp";
 import { validateEncryptionKey } from "../lib/crypto.ts";
-import { initDb } from "../lib/db.ts";
+import { initDb } from "../lib/db/index.ts";
 import { handleRequest } from "../server.ts";
 
 // biome-ignore lint/suspicious/noConsole: Edge script logging
