+++ src/index.ts
@@ -4,7 +4,7 @@
 
 import { handleRequest } from "#routes";
 import { validateEncryptionKey } from "./lib/crypto.ts";
-import { initDb } from "./lib/db.ts";
+import { initDb } from "./lib/db/index.ts";
 
 const startServer = async (port = 3000): Promise<void> => {
   validateEncryptionKey();
