--- src/lib/db.ts
@@ -1,570 +1,0 @@
-/**
- * Database module for ticket reservation system
- * Uses libsql for SQLite-compatible storage
- */
-
-import { type Client, createClient, type InValue } from "@libsql/client";
-import { lazyRef } from "#fp";
-import { decrypt, encrypt, hashPassword, verifyPassword } from "./crypto.ts";
-import type {
-  Attendee,
-  Event,
-  EventWithCount,
-  Session,
-  Settings,
-} from "./types.ts";
-
-const createDbClient = (): Client => {
-  const url = process.env.DB_URL;
-  if (!url) {
-    throw new Error("DB_URL environment variable is required");
-  }
-  return createClient({
-    url,
-    authToken: process.env.DB_TOKEN,
-  });
-};
-
-const [dbGetter, dbSetter] = lazyRef(createDbClient);
-
-/** Query single row, returning null if not found */
-const queryOne = async <T>(sql: string, args: InValue[]): Promise<T | null> => {
-  const result = await getDb().execute({ sql, args });
-  return result.rows.length === 0 ? null : (result.rows[0] as unknown as T);
-};
-
-/** Execute delete/update by field */
-const executeByField = async (
-  table: string,
-  field: string,
-  value: InValue,
-): Promise<void> => {
-  await getDb().execute({
-    sql: `DELETE FROM ${table} WHERE ${field} = ?`,
-    args: [value],
-  });
-};
-
-/**
- * Get or create database client
- */
-export const getDb = (): Client => dbGetter();
-
-/**
- * Set database client (for testing)
- */
-export const setDb = (client: Client | null): void => dbSetter(client);
-
-/**
- * Initialize database tables
- */
-export const initDb = async (): Promise<void> => {
-  const client = getDb();
-
-  await client.execute(`
-    CREATE TABLE IF NOT EXISTS settings (
-      key TEXT PRIMARY KEY,
-      value TEXT NOT NULL
-    )
-  `);
-
-  await client.execute(`
-    CREATE TABLE IF NOT EXISTS events (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      created TEXT NOT NULL,
-      name TEXT NOT NULL,
-      description TEXT NOT NULL,
-      max_attendees INTEGER NOT NULL,
-      thank_you_url TEXT NOT NULL,
-      unit_price INTEGER
-    )
-  `);
-
-  // Migration: add unit_price column if it doesn't exist (for existing databases)
-  try {
-    await client.execute("ALTER TABLE events ADD COLUMN unit_price INTEGER");
-  } catch {
-    // Column already exists, ignore error
-  }
-
-  await client.execute(`
-    CREATE TABLE IF NOT EXISTS attendees (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      event_id INTEGER NOT NULL,
-      name TEXT NOT NULL,
-      email TEXT NOT NULL,
-      created TEXT NOT NULL,
-      stripe_payment_id TEXT,
-      FOREIGN KEY (event_id) REFERENCES events(id)
-    )
-  `);
-
-  // Migration: add stripe_payment_id column if it doesn't exist (for existing databases)
-  try {
-    await client.execute(
-      "ALTER TABLE attendees ADD COLUMN stripe_payment_id TEXT",
-    );
-  } catch {
-    // Column already exists, ignore error
-  }
-
-  await client.execute(`
-    CREATE TABLE IF NOT EXISTS sessions (
-      token TEXT PRIMARY KEY,
-      csrf_token TEXT NOT NULL,
-      expires INTEGER NOT NULL
-    )
-  `);
-
-  // Migration: add csrf_token column if it doesn't exist (for existing databases)
-  try {
-    await client.execute(
-      "ALTER TABLE sessions ADD COLUMN csrf_token TEXT NOT NULL DEFAULT ''",
-    );
-  } catch {
-    // Column already exists, ignore error
-  }
-
-  await client.execute(`
-    CREATE TABLE IF NOT EXISTS login_attempts (
-      ip TEXT PRIMARY KEY,
-      attempts INTEGER NOT NULL DEFAULT 0,
-      locked_until INTEGER
-    )
-  `);
-};
-
-/**
- * Get a setting value
- */
-export const getSetting = async (key: string): Promise<string | null> => {
-  const result = await getDb().execute({
-    sql: "SELECT value FROM settings WHERE key = ?",
-    args: [key],
-  });
-  if (result.rows.length === 0) return null;
-  return (result.rows[0] as unknown as Settings).value;
-};
-
-/**
- * Set a setting value
- */
-export const setSetting = async (key: string, value: string): Promise<void> => {
-  await getDb().execute({
-    sql: "INSERT OR REPLACE INTO settings (key, value) VALUES (?, ?)",
-    args: [key, value],
-  });
-};
-
-/**
- * Setting keys for configuration
- */
-export const CONFIG_KEYS = {
-  ADMIN_PASSWORD: "admin_password",
-  STRIPE_KEY: "stripe_key",
-  CURRENCY_CODE: "currency_code",
-  SETUP_COMPLETE: "setup_complete",
-} as const;
-
-/**
- * Check if initial setup has been completed
- */
-export const isSetupComplete = async (): Promise<boolean> => {
-  const value = await getSetting(CONFIG_KEYS.SETUP_COMPLETE);
-  return value === "true";
-};
-
-/**
- * Complete initial setup by storing all configuration
- * Passwords are hashed using scrypt before storage
- * Sensitive values are encrypted at rest
- */
-export const completeSetup = async (
-  adminPassword: string,
-  stripeSecretKey: string | null,
-  currencyCode: string,
-): Promise<void> => {
-  const hashedPassword = await hashPassword(adminPassword);
-  const encryptedHash = await encrypt(hashedPassword);
-  await setSetting(CONFIG_KEYS.ADMIN_PASSWORD, encryptedHash);
-  if (stripeSecretKey) {
-    const encryptedKey = await encrypt(stripeSecretKey);
-    await setSetting(CONFIG_KEYS.STRIPE_KEY, encryptedKey);
-  }
-  await setSetting(CONFIG_KEYS.CURRENCY_CODE, currencyCode);
-  await setSetting(CONFIG_KEYS.SETUP_COMPLETE, "true");
-};
-
-/**
- * Get Stripe secret key from database (decrypted)
- */
-export const getStripeSecretKeyFromDb = async (): Promise<string | null> => {
-  const value = await getSetting(CONFIG_KEYS.STRIPE_KEY);
-  if (!value) return null;
-  return decrypt(value);
-};
-
-/**
- * Check if a Stripe key has been configured
- */
-export const hasStripeKey = async (): Promise<boolean> => {
-  const value = await getSetting(CONFIG_KEYS.STRIPE_KEY);
-  return value !== null;
-};
-
-/**
- * Update Stripe secret key (encrypted at rest)
- */
-export const updateStripeKey = async (
-  stripeSecretKey: string,
-): Promise<void> => {
-  const encryptedKey = await encrypt(stripeSecretKey);
-  await setSetting(CONFIG_KEYS.STRIPE_KEY, encryptedKey);
-};
-
-/**
- * Get currency code from database
- */
-export const getCurrencyCodeFromDb = async (): Promise<string> => {
-  const value = await getSetting(CONFIG_KEYS.CURRENCY_CODE);
-  return value || "GBP";
-};
-
-/**
- * Get admin password hash from database (decrypted)
- * Returns null if setup hasn't been completed
- */
-export const getAdminPasswordFromDb = async (): Promise<string | null> => {
-  const value = await getSetting(CONFIG_KEYS.ADMIN_PASSWORD);
-  if (!value) return null;
-  return decrypt(value);
-};
-
-/**
- * Verify admin password using constant-time comparison
- * Checks the database-stored password hash only
- */
-export const verifyAdminPassword = async (
-  password: string,
-): Promise<boolean> => {
-  const storedHash = await getAdminPasswordFromDb();
-  if (storedHash === null) return false;
-  return verifyPassword(password, storedHash);
-};
-
-/**
- * Update admin password and invalidate all existing sessions
- * Passwords are hashed using scrypt before storage and encrypted at rest
- */
-export const updateAdminPassword = async (
-  newPassword: string,
-): Promise<void> => {
-  const hashedPassword = await hashPassword(newPassword);
-  const encryptedHash = await encrypt(hashedPassword);
-  await setSetting(CONFIG_KEYS.ADMIN_PASSWORD, encryptedHash);
-  await deleteAllSessions();
-};
-
-/** Event input fields for create/update */
-export type EventInput = {
-  name: string;
-  description: string;
-  maxAttendees: number;
-  thankYouUrl: string;
-  unitPrice?: number | null;
-};
-
-/**
- * Create a new event
- */
-export const createEvent = async (e: EventInput): Promise<Event> => {
-  const created = new Date().toISOString();
-  const args: InValue[] = [
-    created,
-    e.name,
-    e.description,
-    e.maxAttendees,
-    e.thankYouUrl,
-    e.unitPrice ?? null,
-  ];
-  const result = await getDb().execute({
-    sql: `INSERT INTO events (created, name, description, max_attendees, thank_you_url, unit_price)
-          VALUES (?, ?, ?, ?, ?, ?)`,
-    args,
-  });
-  return {
-    id: Number(result.lastInsertRowid),
-    created,
-    name: e.name,
-    description: e.description,
-    max_attendees: e.maxAttendees,
-    thank_you_url: e.thankYouUrl,
-    unit_price: e.unitPrice ?? null,
-  };
-};
-
-/**
- * Update an existing event
- */
-export const updateEvent = async (
-  id: number,
-  e: EventInput,
-): Promise<Event | null> => {
-  const args: InValue[] = [
-    e.name,
-    e.description,
-    e.maxAttendees,
-    e.thankYouUrl,
-    e.unitPrice ?? null,
-    id,
-  ];
-  const result = await getDb().execute({
-    sql: `UPDATE events SET name = ?, description = ?, max_attendees = ?, thank_you_url = ?, unit_price = ?
-          WHERE id = ?`,
-    args,
-  });
-  if (result.rowsAffected === 0) return null;
-  return getEvent(id);
-};
-
-/**
- * Get all events with attendee counts
- */
-export const getAllEvents = async (): Promise<EventWithCount[]> => {
-  const result = await getDb().execute(`
-    SELECT e.*, COUNT(a.id) as attendee_count
-    FROM events e
-    LEFT JOIN attendees a ON e.id = a.event_id
-    GROUP BY e.id
-    ORDER BY e.created DESC
-  `);
-  return result.rows as unknown as EventWithCount[];
-};
-
-/**
- * Get a single event by ID
- */
-export const getEvent = async (id: number): Promise<Event | null> =>
-  queryOne<Event>("SELECT * FROM events WHERE id = ?", [id]);
-
-/**
- * Get event with attendee count
- */
-export const getEventWithCount = async (
-  id: number,
-): Promise<EventWithCount | null> =>
-  queryOne<EventWithCount>(
-    `SELECT e.*, COUNT(a.id) as attendee_count
-     FROM events e
-     LEFT JOIN attendees a ON e.id = a.event_id
-     WHERE e.id = ?
-     GROUP BY e.id`,
-    [id],
-  );
-
-/**
- * Decrypt attendee fields
- */
-const decryptAttendee = async (row: Attendee): Promise<Attendee> => {
-  const name = await decrypt(row.name);
-  const email = await decrypt(row.email);
-  const stripe_payment_id = row.stripe_payment_id
-    ? await decrypt(row.stripe_payment_id)
-    : null;
-  return { ...row, name, email, stripe_payment_id };
-};
-
-/**
- * Get attendees for an event (decrypted)
- */
-export const getAttendees = async (eventId: number): Promise<Attendee[]> => {
-  const result = await getDb().execute({
-    sql: "SELECT * FROM attendees WHERE event_id = ? ORDER BY created DESC",
-    args: [eventId],
-  });
-  const rows = result.rows as unknown as Attendee[];
-  const decrypted: Attendee[] = [];
-  for (const row of rows) {
-    decrypted.push(await decryptAttendee(row));
-  }
-  return decrypted;
-};
-
-/**
- * Create a new attendee (reserve a ticket)
- * Sensitive fields (name, email, stripe_payment_id) are encrypted at rest
- */
-export const createAttendee = async (
-  eventId: number,
-  name: string,
-  email: string,
-  stripePaymentId: string | null = null,
-): Promise<Attendee> => {
-  const created = new Date().toISOString();
-  const encryptedName = await encrypt(name);
-  const encryptedEmail = await encrypt(email);
-  const encryptedPaymentId = stripePaymentId
-    ? await encrypt(stripePaymentId)
-    : null;
-  const result = await getDb().execute({
-    sql: "INSERT INTO attendees (event_id, name, email, created, stripe_payment_id) VALUES (?, ?, ?, ?, ?)",
-    args: [eventId, encryptedName, encryptedEmail, created, encryptedPaymentId],
-  });
-  return {
-    id: Number(result.lastInsertRowid),
-    event_id: eventId,
-    name,
-    email,
-    created,
-    stripe_payment_id: stripePaymentId,
-  };
-};
-
-/**
- * Get an attendee by ID (decrypted)
- */
-export const getAttendee = async (id: number): Promise<Attendee | null> => {
-  const row = await queryOne<Attendee>("SELECT * FROM attendees WHERE id = ?", [
-    id,
-  ]);
-  return row ? decryptAttendee(row) : null;
-};
-
-/**
- * Update attendee's Stripe payment ID (encrypted at rest)
- */
-export const updateAttendeePayment = async (
-  attendeeId: number,
-  stripePaymentId: string,
-): Promise<void> => {
-  const encryptedPaymentId = await encrypt(stripePaymentId);
-  await getDb().execute({
-    sql: "UPDATE attendees SET stripe_payment_id = ? WHERE id = ?",
-    args: [encryptedPaymentId, attendeeId],
-  });
-};
-
-/**
- * Delete an attendee (for cleanup on payment failure)
- */
-export const deleteAttendee = async (attendeeId: number): Promise<void> =>
-  executeByField("attendees", "id", attendeeId);
-
-/**
- * Check if event has available spots
- */
-export const hasAvailableSpots = async (eventId: number): Promise<boolean> => {
-  const event = await getEventWithCount(eventId);
-  if (!event) return false;
-  return event.attendee_count < event.max_attendees;
-};
-
-/**
- * Create a new session with CSRF token
- */
-export const createSession = async (
-  token: string,
-  csrfToken: string,
-  expires: number,
-): Promise<void> => {
-  await getDb().execute({
-    sql: "INSERT INTO sessions (token, csrf_token, expires) VALUES (?, ?, ?)",
-    args: [token, csrfToken, expires],
-  });
-};
-
-/**
- * Get a session by token
- */
-export const getSession = async (token: string): Promise<Session | null> =>
-  queryOne<Session>(
-    "SELECT token, csrf_token, expires FROM sessions WHERE token = ?",
-    [token],
-  );
-
-/**
- * Delete a session by token
- */
-export const deleteSession = async (token: string): Promise<void> =>
-  executeByField("sessions", "token", token);
-
-/**
- * Delete all expired sessions
- */
-export const deleteExpiredSessions = async (): Promise<void> => {
-  await getDb().execute({
-    sql: "DELETE FROM sessions WHERE expires < ?",
-    args: [Date.now()],
-  });
-};
-
-/**
- * Delete all sessions (used when password is changed)
- */
-export const deleteAllSessions = async (): Promise<void> => {
-  await getDb().execute("DELETE FROM sessions");
-};
-
-/**
- * Rate limiting constants
- */
-const MAX_LOGIN_ATTEMPTS = 5;
-const LOCKOUT_DURATION_MS = 15 * 60 * 1000; // 15 minutes
-
-/** Query login attempts for an IP */
-const getLoginAttempts = async (ip: string) =>
-  queryOne<{ attempts: number; locked_until: number | null }>(
-    "SELECT attempts, locked_until FROM login_attempts WHERE ip = ?",
-    [ip],
-  );
-
-/**
- * Check if IP is rate limited for login
- */
-export const isLoginRateLimited = async (ip: string): Promise<boolean> => {
-  const row = await getLoginAttempts(ip);
-  if (!row) return false;
-
-  // Check if currently locked out
-  if (row.locked_until && row.locked_until > Date.now()) {
-    return true;
-  }
-
-  // If lockout expired, reset
-  if (row.locked_until && row.locked_until <= Date.now()) {
-    await executeByField("login_attempts", "ip", ip);
-    return false;
-  }
-
-  return false;
-};
-
-/**
- * Record a failed login attempt
- * Returns true if the account is now locked
- */
-export const recordFailedLogin = async (ip: string): Promise<boolean> => {
-  const row = await getLoginAttempts(ip);
-  const newAttempts = (row?.attempts ?? 0) + 1;
-
-  if (newAttempts >= MAX_LOGIN_ATTEMPTS) {
-    const lockedUntil = Date.now() + LOCKOUT_DURATION_MS;
-    await getDb().execute({
-      sql: "INSERT OR REPLACE INTO login_attempts (ip, attempts, locked_until) VALUES (?, ?, ?)",
-      args: [ip, newAttempts, lockedUntil],
-    });
-    return true;
-  }
-
-  await getDb().execute({
-    sql: "INSERT OR REPLACE INTO login_attempts (ip, attempts, locked_until) VALUES (?, ?, NULL)",
-    args: [ip, newAttempts],
-  });
-  return false;
-};
-
-/**
- * Clear login attempts for an IP (on successful login)
- */
-export const clearLoginAttempts = async (ip: string): Promise<void> =>
-  executeByField("login_attempts", "ip", ip);
