+++ src/lib/db/index.ts
@@ -1,0 +1,58 @@
+/**
+ * Database module for ticket reservation system
+ * Uses libsql for SQLite-compatible storage
+ */
+
+// Attendees
+export {
+  createAttendee,
+  deleteAttendee,
+  getAttendee,
+  getAttendees,
+  hasAvailableSpots,
+  updateAttendeePayment,
+} from "./attendees.ts";
+// Client
+export { getDb, setDb } from "./client.ts";
+export type { EventInput } from "./events.ts";
+
+// Events
+export {
+  createEvent,
+  getAllEvents,
+  getEvent,
+  getEventWithCount,
+  updateEvent,
+} from "./events.ts";
+// Login attempts
+export {
+  clearLoginAttempts,
+  isLoginRateLimited,
+  recordFailedLogin,
+} from "./login-attempts.ts";
+// Migrations
+export { initDb } from "./migrations/index.ts";
+
+// Sessions
+export {
+  createSession,
+  deleteAllSessions,
+  deleteExpiredSessions,
+  deleteSession,
+  getSession,
+} from "./sessions.ts";
+// Settings
+export {
+  CONFIG_KEYS,
+  completeSetup,
+  getAdminPasswordFromDb,
+  getCurrencyCodeFromDb,
+  getSetting,
+  getStripeSecretKeyFromDb,
+  hasStripeKey,
+  isSetupComplete,
+  setSetting,
+  updateAdminPassword,
+  updateStripeKey,
+  verifyAdminPassword,
+} from "./settings.ts";
