+++ test/lib/config.test.ts
@@ -9,7 +9,7 @@ import {
   isSetupComplete,
 } from "#lib/config.ts";
 import { encrypt } from "#lib/crypto.ts";
-import { completeSetup, setSetting } from "#lib/db.ts";
+import { completeSetup, setSetting } from "#lib/db";
 import { createTestDb, resetDb } from "#test-utils";
 
 describe("config", () => {
