+++ test/lib/stripe.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 import { encrypt } from "#lib/crypto.ts";
-import { setSetting } from "#lib/db.ts";
+import { setSetting } from "#lib/db";
 import {
   createCheckoutSession,
   formatPrice,
