+++ test/lib/test-utils.test.ts
@@ -16,7 +16,7 @@ describe("test-utils", () => {
   describe("createTestDb", () => {
     test("creates an in-memory database that can execute queries", async () => {
       await createTestDb();
-      const { getDb } = await import("#lib/db.ts");
+      const { getDb } = await import("#lib/db");
       const result = await getDb().execute("SELECT 1 as test");
       expect(result.rows.length).toBe(1);
       expect(result.columns).toContain("test");
@@ -26,7 +26,7 @@ describe("test-utils", () => {
   describe("resetDb", () => {
     test("resets database so next getDb creates fresh connection", async () => {
       await createTestDb();
-      const { getDb, setDb } = await import("#lib/db.ts");
+      const { getDb, setDb } = await import("#lib/db");
       const firstDb = getDb();
       resetDb();
       setDb(null);
